feat: added Matrix for more type safety
Psalm does not like standard PHP arrays, especially as PHP does not
discern between dictionaries and lists.
The usual method of using a twodimensional array will result in a lot
of MixedAssignment and so on by Psalm; Matrix encapsulates the handling
of the cell data of TerminalTableModel if you so wish.